Conflicts:
	src/components/dbus/codegen/make_HMIRequests_class.py

+++ src/components/utils/src/file_system.cc
@@ -67,8 +67,8 @@ int64_t file_system::FileSize(const std::string &path) {
   return 0;
 }
 
-uint32_t file_system::DirectorySize(const std::string& path) {
-  uint32_t size = 0;
+size_t file_system::DirectorySize(const std::string& path) {
+  size_t size = 0;
   int32_t return_code = 0;
   DIR* directory = NULL;
